refactor: extract `JsonResponseTrait`
